refactor(internal/cookies): extract cookie manager from pkg/kratos
Move AuthCookieManager and Encrypt from pkg/kratos into a new
internal/cookies package so they can be shared by pkg/tenants
without creating a circular dependency.
